refactor(schemas): migrate Phase 38 endpoint schemas — analytics_bug_prediction (#6042)
Move Pydantic classes and supporting enums from analytics_bug_prediction.py
into schemas_analytics.py.

- RiskLevel enum → schemas_analytics.py
- RiskFactor enum → schemas_analytics.py
- FileRisk → schemas_analytics.py
- PredictionResult → schemas_analytics.py

Endpoint file now imports RiskLevel/RiskFactor (still used by RISK_WEIGHTS
and PREVENTION_TIPS dicts). FileRisk/PredictionResult were defined but
never referenced in the endpoint, so they're left available in the schema
module for future use.
